[refactor] #112 CQS 패턴에 따라 분리
