+++ src/main/java/DGU_AI_LAB/admin_be/domain/requests/service/RequestQueryService.java
@@ -1,0 +1,59 @@
+package DGU_AI_LAB.admin_be.domain.requests.service;
+
+import DGU_AI_LAB.admin_be.domain.containerImage.repository.ContainerImageRepository;
+import DGU_AI_LAB.admin_be.domain.groups.repository.GroupRepository;
+import DGU_AI_LAB.admin_be.domain.requests.dto.response.ContainerInfoDTO;
+import DGU_AI_LAB.admin_be.domain.requests.dto.response.ResourceUsageDTO;
+import DGU_AI_LAB.admin_be.domain.requests.dto.response.SaveRequestResponseDTO;
+import DGU_AI_LAB.admin_be.domain.requests.entity.Status;
+import DGU_AI_LAB.admin_be.domain.requests.repository.ChangeRequestRepository;
+import DGU_AI_LAB.admin_be.domain.requests.repository.RequestRepository;
+import DGU_AI_LAB.admin_be.domain.resourceGroups.repository.ResourceGroupRepository;
+import DGU_AI_LAB.admin_be.domain.users.repository.UserRepository;
+import DGU_AI_LAB.admin_be.error.ErrorCode;
+import DGU_AI_LAB.admin_be.error.exception.BusinessException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class RequestQueryService {
+
+    private final RequestRepository requestRepository;
+    private final UserRepository userRepository;
+
+    /** 내 신청 목록 */
+    @Transactional(readOnly = true)
+    public List<SaveRequestResponseDTO> getRequestsByUserId(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND);
+        }
+        return requestRepository.findAllByUser_UserId(userId).stream()
+                .map(SaveRequestResponseDTO::fromEntity)
+                .toList();
+    }
+
+    /** 승인 완료 자원 사용량 */
+    @Transactional(readOnly = true)
+    public List<ResourceUsageDTO> getAllFulfilledResourceUsage() {
+        return requestRepository.findAllByStatus(Status.FULFILLED).stream()
+                .map(ResourceUsageDTO::fromEntity)
+                .toList();
+    }
+
+    /** 활성 컨테이너 */
+    @Transactional(readOnly = true)
+    public List<ContainerInfoDTO> getActiveContainers() {
+        return requestRepository.findAllByStatus(Status.FULFILLED).stream()
+                .map(ContainerInfoDTO::fromEntity)
+                .toList();
+    }
+
+}
